requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
Django==4.2.4
frozenlist==1.3.3
greenlet==3.1.1
idna==3.4
loader==2017.9.11
magic-filter==1.0.12
multidict==6.0.4
psycopg2==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.0.0
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.35
sqlparse==0.4.4
telebot==0.0.4
TelegramBotAPI==0.3.2
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.14
yarl==1.8.2
